Create UI manager and TextMesh to appear damage cost in the screen. Clean up game state to share tasks on the UI manager

+++ Engine/JimmyGodEngine/Inc/AgentComponent.h
@@ -49,7 +49,7 @@ namespace JimmyGod
 
 		float mMovementSpeed = 0.0f;
 		float mMass = 0.0f;
-		float mArea = 0.0f;
+		int mArea = 0;
 		std::vector<Math::Vector3> mPath;
 
 	private:
